Custom ttf + small fixes
-Created new ttf file for use within the app "Roboto-Runicode-Regular.ttf" that contains all 256 characters used in the final encryption.
-Small fixes
-New sample video

+++ ColourCodes/Assets/Code/Decrypt.cs
@@ -17,7 +17,7 @@ namespace Evernorth.ColourCodes
         public Dictionary<char, Vector3Int> CharToColorKey;
         public string charactersS;
 
-        public int[] iSeed;
+        public byte[] iSeed;
 
         public Vector3Int[] colArray;
         public string eStringData;
